backend/requirements.txt
# Flask and extensions
flask==2.2.3
flask-cors==3.0.10
Werkzeug==2.2.3
Jinja2==3.1.2

# AWS SDK
boto3==1.26.135
botocore==1.29.135

# Environment variables
python-dotenv==1.0.0

# Authentication
pyjwt==2.6.0

# Utilities
requests==2.30.0
python-dateutil==2.8.2

# File processing
pyPDF2==3.0.1
python-docx==0.8.11

# Logging and monitoring
watchtower==3.0.1

# Testing
pytest==7.3.1